MAKE `ScriptsTestCase` that copies required scripts to base_path directory prior to executing tests
